+++ core/modules/modsplit.class.php
@@ -64,6 +64,9 @@ class modsplit extends DolibarrModules
 		$this->description = "Description of module split";
 		// Possible values for version are: 'development', 'experimental' or version
 		$this->version = '2.0.3';
+		// Url to the file with your last numberversion of this module
+		require_once __DIR__ . '/../../class/techatm.class.php';
+		$this->url_last_version = \split\TechATM::getLastModuleVersionUrl($this);
 		// Key used in llx_const table to save module status enabled/disabled
 		// (where MYMODULE is value of property name of module in uppercase)
 		$this->const_name = 'MAIN_MODULE_' . strtoupper($this->name);
@@ -107,7 +110,7 @@ class modsplit extends DolibarrModules
 
 		// Config pages. Put here list of php pages
 		// stored into split/admin directory, used to setup module.
-		//$this->config_page_url = array("split_setup.php@split");
+		$this->config_page_url = array("about.php@split");
 
 		// Dependencies
 		// List of modules id that must be enabled if this module is enabled